streamer: get the streamer server its own file
