+++ scripts/planktoscope/streamer.py
@@ -1,0 +1,87 @@
+################################################################################
+# Streaming PiCamera over server
+################################################################################
+import io
+import socketserver
+import http
+import multiprocessing
+
+
+################################################################################
+# Classes for the PiCamera Streaming
+################################################################################
+class StreamingOutput(object):
+    def __init__(self):
+        self.frame = None
+        self.buffer = io.BytesIO()
+        self.condition = multiprocessing.Condition()
+
+    def write(self, buf):
+        if buf.startswith(b"\xff\xd8"):
+            # New frame, copy the existing buffer's content and notify all
+            # clients it's available
+            self.buffer.truncate()
+            with self.condition:
+                self.frame = self.buffer.getvalue()
+                self.condition.notify_all()
+            self.buffer.seek(0)
+        return self.buffer.write(buf)
+
+
+class StreamingHandler(http.server.BaseHTTPRequestHandler):
+    # Webpage content containing the PiCamera Streaming
+    PAGE = """\
+    <html>
+    <head>
+    <title>PlanktonScope v2 | PiCamera Streaming</title>
+    </head>
+    <body>
+    <img src="stream.mjpg" width="100%" height="100%" />
+    </body>
+    </html>
+    """
+
+    def do_GET(self):
+        if self.path == "/":
+            self.send_response(301)
+            self.send_header("Location", "/index.html")
+            self.end_headers()
+        elif self.path == "/index.html":
+            content = self.PAGE.encode("utf-8")
+            self.send_response(200)
+            self.send_header("Content-Type", "text/html")
+            self.send_header("Content-Length", len(content))
+            self.end_headers()
+            self.wfile.write(content)
+        elif self.path == "/stream.mjpg":
+            self.send_response(200)
+            self.send_header("Age", 0)
+            self.send_header("Cache-Control", "no-cache, private")
+            self.send_header("Pragma", "no-cache")
+            self.send_header(
+                "Content-Type", "multipart/x-mixed-replace; boundary=FRAME"
+            )
+            self.end_headers()
+            try:
+                while True:
+                    with output.condition:
+                        output.condition.wait()
+                        frame = output.frame
+                    self.wfile.write(b"--FRAME\r\n")
+                    self.send_header("Content-Type", "image/jpeg")
+                    self.send_header("Content-Length", len(frame))
+                    self.end_headers()
+                    self.wfile.write(frame)
+                    self.wfile.write(b"\r\n")
+            except Exception as e:
+                logger.warning(
+                    "Removed streaming client %s: %s", self.client_address, str(e)
+                )
+        else:
+            self.send_error(404)
+            self.end_headers()
+
+
+class StreamingServer(socketserver.ThreadingMixIn, http.server.HTTPServer):
+    allow_reuse_address = True
+    daemon_threads = True
